Update elevator / elbow controls
Add soft stop limit for elbow based on the elevator position. The elevator also uses the relative encoder.

Until the elevator touches the down stop, the elevator position is "uncalibrated" and will only honor the hard limit switches. The soft limits will be disabled until the bottom limit switch is hit.

Also the capability to kick the elbow out or in based on elevator position commands. This is not wired to the controller yet.

MotorIO interface was also updated to support either absolute or external encoder. Elevator is the only system the uses external, all other motors are absolute, including the elbow joint.