Cleanup and put elbow control in the default elevator command
Remove dead/commented code.

Make dE application constants.

Todo still need to allow full range of movement and move upper limits.

Put elbow control on the same command as the elevator, so the commands do not fight with each other. This will allow the elbow and elevator to be controlled at the same time.