Attempt to fix initial robot pose
At teleop init, set the pose based on alliance.
This will only happen if the vision subsystem has not provided a reliable position estimate prior to teleop.

Note the POSE is field relative with blue alliance driver right corner as the origin, a.k.a. the nominal field coordinate system.